compartmentalise logic for garden screen
